TenderPlatform/requirements.txt
txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.0
python-multipart==0.0.12
jinja2==3.1.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
aiofiles==24.1.0
weasyprint==60.2      # для генерации PDF
# psycopg2-binary==2.9.9  # для PostgreSQL (раскомментировать при переходе)